Get conditions and processes to accede memberships

The function allows to get information on the conditions for parties to enter a treaty. It could be limited to a entity area, a certain domain, or on State party nominations for new members. The function also allows to get information on the process by which a new member has to go through in order to accede memberships

Usage

code_memberships(t, title = NULL, memberships = NULL)

Arguments

t

A text variable

title

A title variable. If not declared, NULL by default.

memberships

The "condition" or "process" for parties to accede a treaty. If not declared, returns the "process".

Value

Either the conditions to be part of the treaty or the different process steps to become a member.

Examples

# \donttest{
m <- manyenviron::texts$AGR_TXT[100:300,]
code_memberships(m$Text, m$Title, memberships = "condition")
#> ℹ Please make sure JavaScript is installed.
#>   [1] "entity: River Mino + domain: fishing"                       
#>   [2] "entity: Lower Lake Constance + domain: fishing"             
#>   [3] "NA"                                                         
#>   [4] "NA"                                                         
#>   [5] "domain: fishing"                                            
#>   [6] "domain: waste"                                              
#>   [7] "entity: Of + domain: waste"                                 
#>   [8] "open + domain: environment"                                 
#>   [9] "domain: fishing"                                            
#>  [10] "domain: fishing"                                            
#>  [11] "domain: fishing"                                            
#>  [12] "domain: fishing"                                            
#>  [13] "open + domain: agriculture"                                 
#>  [14] "domain: fishing"                                            
#>  [15] "domain: waste"                                              
#>  [16] "domain: fishing"                                            
#>  [17] "domain: fishing"                                            
#>  [18] "NA"                                                         
#>  [19] "domain: fishing"                                            
#>  [20] "domain: waste"                                              
#>  [21] "NA"                                                         
#>  [22] "open"                                                       
#>  [23] "NA"                                                         
#>  [24] "domain: fishing"                                            
#>  [25] "NA"                                                         
#>  [26] "NA"                                                         
#>  [27] "domain: fishing"                                            
#>  [28] "domain: fishing"                                            
#>  [29] "domain: fishing"                                            
#>  [30] "domain: fishing"                                            
#>  [31] "domain: fishing"                                            
#>  [32] "domain: fishing"                                            
#>  [33] "open + domain: territorial boundaries"                      
#>  [34] "entity: North Atlantic Coast + domain: fishing"             
#>  [35] "NA"                                                         
#>  [36] "domain: waste"                                              
#>  [37] "NA"                                                         
#>  [38] "open"                                                       
#>  [39] "domain: fishing"                                            
#>  [40] "NA"                                                         
#>  [41] "domain: trade"                                              
#>  [42] "domain: defense"                                            
#>  [43] "NA"                                                         
#>  [44] "domain: fishing"                                            
#>  [45] "NA"                                                         
#>  [46] "domain: fishing"                                            
#>  [47] "open"                                                       
#>  [48] "open + entity: Congo River Basin"                           
#>  [49] "open"                                                       
#>  [50] "NA"                                                         
#>  [51] "domain: fishing"                                            
#>  [52] "open"                                                       
#>  [53] "domain: fishing"                                            
#>  [54] "domain: fishing"                                            
#>  [55] "open"                                                       
#>  [56] "domain: territorial boundaries"                             
#>  [57] "domain: law enforcement"                                    
#>  [58] "entity: Videa River + domain: fishing"                      
#>  [59] "domain: fishing"                                            
#>  [60] "domain: fishing"                                            
#>  [61] "domain: fishing"                                            
#>  [62] "domain: fishing"                                            
#>  [63] "open + domain: fishing"                                     
#>  [64] "open + entity: Arctic Ocean + domain: fishing"              
#>  [65] "open + domain: fishing"                                     
#>  [66] "open + domain: fishing"                                     
#>  [67] "domain: territorial boundaries"                             
#>  [68] "open + domain: fishing"                                     
#>  [69] "domain: fishing"                                            
#>  [70] "open + domain: energy"                                      
#>  [71] "open + domain: energy"                                      
#>  [72] "NA"                                                         
#>  [73] "open + domain: health"                                      
#>  [74] "domain: fishing"                                            
#>  [75] "NA"                                                         
#>  [76] "NA"                                                         
#>  [77] "NA"                                                         
#>  [78] "NA"                                                         
#>  [79] "domain: territorial boundaries"                             
#>  [80] "open + domain: territorial boundaries"                      
#>  [81] "domain: health"                                             
#>  [82] "domain: fishing"                                            
#>  [83] "NA"                                                         
#>  [84] "NA"                                                         
#>  [85] "NA"                                                         
#>  [86] "domain: fishing"                                            
#>  [87] "domain: fishing"                                            
#>  [88] "open + entity: Palestine + domain: research"                
#>  [89] "NA"                                                         
#>  [90] "entity: Kunene River West Africa + domain: energy"          
#>  [91] "NA"                                                         
#>  [92] "domain: energy"                                             
#>  [93] "entity: Caspian Sea + domain: fishing"                      
#>  [94] "NA"                                                         
#>  [95] "domain: fishing"                                            
#>  [96] "domain: fishing"                                            
#>  [97] "domain: fishing"                                            
#>  [98] "domain: agriculture"                                        
#>  [99] "entity: Granville Bay + domain: fishing"                    
#> [100] "open + domain: environment"                                 
#> [101] "domain: law enforcement"                                    
#> [102] "open + entity: Baltic Sea + domain: fishing"                
#> [103] "open + domain: fishing"                                     
#> [104] "domain: fishing"                                            
#> [105] "open + entity: Fraser River + domain: fishing"              
#> [106] "entity: Sverdrup Islands + domain: fishing"                 
#> [107] "domain: defense"                                            
#> [108] "domain: fishing"                                            
#> [109] "NA"                                                         
#> [110] "open + domain: fishing"                                     
#> [111] "NA"                                                         
#> [112] "NA"                                                         
#> [113] "domain: trade"                                              
#> [114] "open"                                                       
#> [115] "domain: fishing"                                            
#> [116] "domain: fishing"                                            
#> [117] "NA"                                                         
#> [118] "open"                                                       
#> [119] "open"                                                       
#> [120] "open"                                                       
#> [121] "open + entity: Lake Ladoga + domain: fishing"               
#> [122] "domain: fishing"                                            
#> [123] "domain: fishing"                                            
#> [124] "NA"                                                         
#> [125] "open + domain: health"                                      
#> [126] "open + domain: agriculture"                                 
#> [127] "open + domain: agriculture"                                 
#> [128] "entity: Trail British Columbia + domain: law enforcement"   
#> [129] "NA"                                                         
#> [130] "entity: Lake Titicaca + domain: fishing"                    
#> [131] "domain: territorial boundaries"                             
#> [132] "domain: environment"                                        
#> [133] "NA"                                                         
#> [134] "domain: fishing"                                            
#> [135] "open + domain: fishing"                                     
#> [136] "open + domain: fishing"                                     
#> [137] "domain: fishing"                                            
#> [138] "open"                                                       
#> [139] "Semi-open + domain: environment"                            
#> [140] "entity: Tana River + domain: fishing"                       
#> [141] "entity: Pasvik River + domain: fishing"                     
#> [142] "domain: health"                                             
#> [143] "open + domain: fishing"                                     
#> [144] "open + domain: fishing"                                     
#> [145] "domain: territorial boundaries"                             
#> [146] "NA"                                                         
#> [147] "domain: fishing"                                            
#> [148] "domain: fishing"                                            
#> [149] "open"                                                       
#> [150] "domain: trade"                                              
#> [151] "NA"                                                         
#> [152] "domain: trade"                                              
#> [153] "open + entity: Niagara River Niagara Falls + domain: energy"
#> [154] "open + domain: environment"                                 
#> [155] "domain: trade"                                              
#> [156] "domain: trade"                                              
#> [157] "open + domain: agriculture"                                 
#> [158] "NA"                                                         
#> [159] "open"                                                       
#> [160] "open + domain: fishing"                                     
#> [161] "entity: Upper Columbia River Basin + domain: trade"         
#> [162] "domain: fishing"                                            
#> [163] "entity: Fort"                                               
#> [164] "entity: Fort"                                               
#> [165] "domain: territorial boundaries"                             
#> [166] "open + domain: fishing"                                     
#> [167] "open + domain: agriculture"                                 
#> [168] "open + domain: fishing"                                     
#> [169] "open + domain: fishing"                                     
#> [170] "open + domain: fishing"                                     
#> [171] "NA"                                                         
#> [172] "domain: fishing"                                            
#> [173] "open + domain: fishing"                                     
#> [174] "entity: River"                                              
#> [175] "domain: fishing"                                            
#> [176] "domain: fishing"                                            
#> [177] "NA"                                                         
#> [178] "domain: trade"                                              
#> [179] "open + domain: fishing"                                     
#> [180] "open + domain: fishing"                                     
#> [181] "domain: environment"                                        
#> [182] "domain: fishing"                                            
#> [183] "domain: law enforcement"                                    
#> [184] "open + domain: research"                                    
#> [185] "domain: fishing"                                            
#> [186] "open + domain: fishing"                                     
#> [187] "domain: territorial boundaries"                             
#> [188] "open"                                                       
#> [189] "domain: fishing"                                            
#> [190] "open + domain: fishing"                                     
#> [191] "entity: Falls + domain: trade"                              
#> [192] "open + domain: fishing"                                     
#> [193] "open + domain: fishing"                                     
#> [194] "open + domain: fishing"                                     
#> [195] "open + domain: fishing"                                     
#> [196] "entity: Tana River + domain: fishing"                       
#> [197] "open + entity: Mediterranean + domain: fishing"             
#> [198] "entity: Trail British Columbia + domain: law enforcement"   
#> [199] "open + entity: Niagara River"                               
#> [200] "open + domain: fishing"                                     
#> [201] "domain: environment"                                        
code_memberships(m$Text, memberships = "process")
#>   [1] "NA"                                     
#>   [2] "NA"                                     
#>   [3] "ratification"                           
#>   [4] "NA"                                     
#>   [5] "NA"                                     
#>   [6] "NA"                                     
#>   [7] "NA"                                     
#>   [8] "ratification + notification"            
#>   [9] "NA"                                     
#>  [10] "NA"                                     
#>  [11] "NA"                                     
#>  [12] "NA"                                     
#>  [13] "NA"                                     
#>  [14] "NA"                                     
#>  [15] "NA"                                     
#>  [16] "NA"                                     
#>  [17] "NA"                                     
#>  [18] "NA"                                     
#>  [19] "NA"                                     
#>  [20] "NA"                                     
#>  [21] "NA"                                     
#>  [22] "ratification"                           
#>  [23] "NA"                                     
#>  [24] "NA"                                     
#>  [25] "NA"                                     
#>  [26] "NA"                                     
#>  [27] "NA"                                     
#>  [28] "NA"                                     
#>  [29] "NA"                                     
#>  [30] "NA"                                     
#>  [31] "NA"                                     
#>  [32] "NA"                                     
#>  [33] "ratification"                           
#>  [34] "NA"                                     
#>  [35] "NA"                                     
#>  [36] "NA"                                     
#>  [37] "NA"                                     
#>  [38] "ratification"                           
#>  [39] "NA"                                     
#>  [40] "NA"                                     
#>  [41] "NA"                                     
#>  [42] "NA"                                     
#>  [43] "NA"                                     
#>  [44] "NA"                                     
#>  [45] "NA"                                     
#>  [46] "NA"                                     
#>  [47] "ratification"                           
#>  [48] "ratification + notification"            
#>  [49] "ratification + notification"            
#>  [50] "NA"                                     
#>  [51] "NA"                                     
#>  [52] "ratification + notification"            
#>  [53] "NA"                                     
#>  [54] "NA"                                     
#>  [55] "ratification + notification"            
#>  [56] "NA"                                     
#>  [57] "NA"                                     
#>  [58] "NA"                                     
#>  [59] "NA"                                     
#>  [60] "NA"                                     
#>  [61] "NA"                                     
#>  [62] "NA"                                     
#>  [63] "ratification + notification"            
#>  [64] "notification"                           
#>  [65] "NA"                                     
#>  [66] "ratification"                           
#>  [67] "NA"                                     
#>  [68] "ratification"                           
#>  [69] "NA"                                     
#>  [70] "signature + ratification"               
#>  [71] "NA"                                     
#>  [72] "NA"                                     
#>  [73] "signature + ratification + notification"
#>  [74] "NA"                                     
#>  [75] "NA"                                     
#>  [76] "NA"                                     
#>  [77] "NA"                                     
#>  [78] "NA"                                     
#>  [79] "NA"                                     
#>  [80] "ratification + notification"            
#>  [81] "ratification + notification"            
#>  [82] "NA"                                     
#>  [83] "NA"                                     
#>  [84] "NA"                                     
#>  [85] "NA"                                     
#>  [86] "NA"                                     
#>  [87] "NA"                                     
#>  [88] "notification"                           
#>  [89] "NA"                                     
#>  [90] "NA"                                     
#>  [91] "NA"                                     
#>  [92] "NA"                                     
#>  [93] "NA"                                     
#>  [94] "NA"                                     
#>  [95] "NA"                                     
#>  [96] "NA"                                     
#>  [97] "NA"                                     
#>  [98] "NA"                                     
#>  [99] "NA"                                     
#> [100] "ratification + notification"            
#> [101] "NA"                                     
#> [102] "ratification"                           
#> [103] "ratification"                           
#> [104] "NA"                                     
#> [105] "ratification"                           
#> [106] "NA"                                     
#> [107] "NA"                                     
#> [108] "NA"                                     
#> [109] "NA"                                     
#> [110] "ratification + notification"            
#> [111] "NA"                                     
#> [112] "NA"                                     
#> [113] "NA"                                     
#> [114] "ratification"                           
#> [115] "NA"                                     
#> [116] "NA"                                     
#> [117] "NA"                                     
#> [118] "ratification"                           
#> [119] "signature + ratification + notification"
#> [120] "NA"                                     
#> [121] "ratification"                           
#> [122] "NA"                                     
#> [123] "NA"                                     
#> [124] "signature + ratification"               
#> [125] "ratification + notification"            
#> [126] "ratification + notification"            
#> [127] "ratification"                           
#> [128] "NA"                                     
#> [129] "NA"                                     
#> [130] "NA"                                     
#> [131] "NA"                                     
#> [132] "NA"                                     
#> [133] "NA"                                     
#> [134] "NA"                                     
#> [135] "ratification + notification"            
#> [136] "ratification + notification"            
#> [137] "NA"                                     
#> [138] "ratification"                           
#> [139] "ratification"                           
#> [140] "NA"                                     
#> [141] "NA"                                     
#> [142] "NA"                                     
#> [143] "ratification + notification"            
#> [144] "NA"                                     
#> [145] "NA"                                     
#> [146] "NA"                                     
#> [147] "NA"                                     
#> [148] "NA"                                     
#> [149] "signature + ratification"               
#> [150] "NA"                                     
#> [151] "ratification"                           
#> [152] "NA"                                     
#> [153] "NA"                                     
#> [154] "ratification"                           
#> [155] "NA"                                     
#> [156] "NA"                                     
#> [157] "ratification + notification"            
#> [158] "NA"                                     
#> [159] "ratification"                           
#> [160] "ratification"                           
#> [161] "NA"                                     
#> [162] "NA"                                     
#> [163] "NA"                                     
#> [164] "NA"                                     
#> [165] "NA"                                     
#> [166] "ratification"                           
#> [167] "notification"                           
#> [168] "signature + ratification"               
#> [169] "ratification"                           
#> [170] "ratification"                           
#> [171] "NA"                                     
#> [172] "signature"                              
#> [173] "signature + ratification"               
#> [174] "NA"                                     
#> [175] "signature + ratification"               
#> [176] "NA"                                     
#> [177] "NA"                                     
#> [178] "NA"                                     
#> [179] "majority vote"                          
#> [180] "majority vote"                          
#> [181] "NA"                                     
#> [182] "NA"                                     
#> [183] "NA"                                     
#> [184] "ratification"                           
#> [185] "NA"                                     
#> [186] "signature + ratification + notification"
#> [187] "NA"                                     
#> [188] "ratification + notification"            
#> [189] "NA"                                     
#> [190] "ratification"                           
#> [191] "NA"                                     
#> [192] "NA"                                     
#> [193] "NA"                                     
#> [194] "NA"                                     
#> [195] "NA"                                     
#> [196] "NA"                                     
#> [197] "ratification  + majority vote"          
#> [198] "NA"                                     
#> [199] "ratification"                           
#> [200] "ratification"                           
#> [201] "NA"                                     
code_memberships()
#> 
#> 
#> Table: Memberships process steps and criterion
#> 
#> Process                                                                                    Criterion                                                                                                       
#> -----------------------------------------------------------------------------------------  ----------------------------------------------------------------------------------------------------------------
#> Majority vote: accession of a new member to the convention is subject to majority voting   Activity: treaty restricted to the exercise of a certain activity                                               
#> Notification: new member states has to notify their consent to the competent body          Geographic: treaty restricted to a certain geographical area                                                    
#> Ratification: requires ratification from the new member                                    Open: treaty open to any new member states within the geographical area and respecting conditions for accession 
#> Signature: requires signature from the new member                                          Semi-open: accession of a new member to the treaty is subject to a vote from the member states                  
#> Unanimity: accession of a new member to the convention is subject to unanimity voting      NA                                                                                                              
# }