Get links from manyID

Treaties that modify, amend, or expand other treaties usually specify so in the title. We use the manyIDs generated by code_agreements() and condense_agreements() to return a data frame of treaty links for a database.

Usage

get_links(database, dataset, treaty_type = "all")

Arguments

database

A Package database

dataset

A Package dataset

treaty_type

The type of treaties to be returned. By default, all treaties are returned. Other options include bilateral or multilateral treaties.

Value

A dataframe of agreements' treatyID and their linkages.

Examples

if (FALSE) {
get_links(database = manyenviron::agreements)
get_links(database = manyenviron::agreements, treaty_type = "multilateral")
get_links(dataset = manyenviron::agreements$IEADB, treaty_type = "bilateral")
samples <- lapply(manyenviron::agreements,
function(x) x[x$Beg > "1991-12-31" & x$Beg < "1993-01-01", ])
migraph::gglineage(get_links(samples))
}